pyproject.toml
# Project
[project]
name = "anncsu-update-action"
version = "0.1.0"
description = "GitHub Action to update ANNCSU db via openapi using anncsu-sdk"
requires-python = ">=3.10"
classifiers = ["Private :: Do Not Upload"]
dependencies = [
  "actions-tools[github]",
  "anncsu-sdk>=0.0.5",
  "pydantic>=2.9.0",
]

[dependency-groups]
dev = [
  "bandit",
  "pyyaml",
  "ruff",
  "tombi",
  "ty",
  "types-PyYAML",
]
test = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "pytest-mock>=3.12.0",
]

# Bandit
# https://bandit.readthedocs.io/en/latest/config.html
[tool.bandit]
exclude_dirs = [".venv", ".pixi"]
# Skip low-severity subprocess warnings for git commands
skips = ["B404", "B603", "B607", "B110"]

# Coverage
# https://coverage.readthedocs.io/en/latest/
[tool.coverage.run]
omit = ["*/.venv/*", "*/__pycache__/*", "*.pyc"]
source = ["src"]

# Ruff
# https://docs.astral.sh/ruff/configuration/
[tool.ruff]
line-length = 119
target-version = "py310"
extend-exclude = [".github"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "Q"]

# Pytest
# https://docs.pytest.org/en/stable/reference/customize.html
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
anncsu-update-action = { path = ".", editable = true }

[tool.pixi.environments]
default = { solve-group = "default" }
dev = { features = ["dev"], solve-group = "default" }
test = { features = ["test"], solve-group = "default" }

[tool.pixi.tasks]
